Moved MultipitchEvaluator to utils subpackage, removed unused datasets from training script, and added colorbar to plot_magnitude function.
