+++ .gitignore
@@ -8,5 +8,3 @@
 
 *.egg
 *.egg-info
-
-*deep_salience
